# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gaza Future Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Story {
  id: number;
  title: string;
  titleAr: string;
  content: string;
  contentAr: string;
  category: string;
  ageGroup: string;
  coverEmoji: string;
  readingTimeMinutes: number;
  isFeatured?: boolean;
  createdAt: string;
}

export interface StoryInput {
  title: string;
  titleAr: string;
  content: string;
  contentAr: string;
  category: string;
  ageGroup: string;
  coverEmoji: string;
  readingTimeMinutes: number;
  isFeatured?: boolean;
}

export interface Podcast {
  id: number;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  category: string;
  durationMinutes: number;
  audioUrl: string;
  coverEmoji: string;
  createdAt: string;
}

export interface PodcastInput {
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  category: string;
  durationMinutes: number;
  audioUrl: string;
  coverEmoji: string;
}

export interface CurriculumInput {
  subject: string;
  topic: string;
  gradeLevel: string;
  language?: string;
}

export interface CurriculumTransformation {
  id: number;
  subject: string;
  topic: string;
  gradeLevel: string;
  storyTitle: string;
  storyContent: string;
  createdAt: string;
}

export interface TutorConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface TutorConversationInput {
  title: string;
}

export interface TutorMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface TutorConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: TutorMessage[];
}

export interface TutorMessageInput {
  content: string;
}

export interface LibraryStats {
  totalStories: number;
  totalPodcasts: number;
  totalTransformations: number;
  featuredCount: number;
}

export type ListStoriesParams = {
category?: string;
ageGroup?: string;
};

export type ListPodcastsParams = {
category?: string;
};