# lib/api-zod/src/generated/types/libraryStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gaza Future Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface LibraryStats {
  totalStories: number;
  totalPodcasts: number;
  totalTransformations: number;
  featuredCount: number;
}