# lib/api-zod/src/generated/types/podcast.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gaza Future Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Podcast {
  id: number;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  category: string;
  durationMinutes: number;
  audioUrl: string;
  coverEmoji: string;
  createdAt: Date;
}