# lib/api-zod/src/generated/types/podcastInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gaza Future Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface PodcastInput {
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  category: string;
  durationMinutes: number;
  audioUrl: string;
  coverEmoji: string;
}