# lib/api-zod/src/generated/types/storyInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gaza Future Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface StoryInput {
  title: string;
  titleAr: string;
  content: string;
  contentAr: string;
  category: string;
  ageGroup: string;
  coverEmoji: string;
  readingTimeMinutes: number;
  isFeatured?: boolean;
}