# lib/api-zod/src/generated/types/tutorConversation.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gaza Future Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface TutorConversation {
  id: number;
  title: string;
  createdAt: Date;
}