# lib/api-zod/src/generated/types/tutorConversationWithMessages.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gaza Future Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { TutorMessage } from './tutorMessage';

export interface TutorConversationWithMessages {
  id: number;
  title: string;
  createdAt: Date;
  messages: TutorMessage[];
}