# lib/api-zod/src/generated/types/tutorMessage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gaza Future Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface TutorMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: Date;
}